pyproject.toml
[tool.poetry]
name = "multidocs"
description = "Generate documentation from multiple sources."
authors = ["Andreas Lutro <anlutro@gmail.com>"]
version = "0.1"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
bleach = "^2.1"
commonmarkextensions = "^0.8"
flask = "^1.0"
flask-oauthlib = "^0.9"
jinja2 = "^2.8"
python-slugify = "^0.12"
pyyaml = "^3.12"
whoosh = "^2.7"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-flask = "*"
pylint = "*"
black = "*"

[tool.poetry.scripts]
multidocs = "multidocs.cli:main"
